+++ C#/.vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\Toll-Calculator-API\\Service"
+  ],
+  "SelectedNode": "\\LICENSE",
+  "PreviewInSolutionExplorer": false
+}
